[Breacking change] move sentinel2 tools to rio_tiler.sentinel2
